backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.0.2
pandas==2.3.3
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
six==1.17.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
tzdata==2026.1
Werkzeug==3.1.8
zipp==3.23.0
